DDF-3499 Pulling current location validation logic into new WktValidator superclass
