+++ catalog/catalog-app/pom.xml
@@ -494,6 +494,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>org.codice.ddf.validator</groupId>
+            <artifactId>catalog-validator-metacardwkt</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>ddf.catalog.transformer</groupId>
             <artifactId>catalog-transformer-overlay</artifactId>
             <version>${project.version}</version>
