+++ catalog/validator/catalog-validator-metacardlocation/src/main/java/org/codice/ddf/validator/metacard/location/MetacardLocationValidator.java
@@ -13,85 +13,13 @@
  */
 package org.codice.ddf.validator.metacard.location;
 
-import com.google.common.collect.ImmutableSet;
-import ddf.catalog.data.Attribute;
-import ddf.catalog.data.Metacard;
-import ddf.catalog.data.impl.AttributeImpl;
 import ddf.catalog.data.types.Core;
-import ddf.catalog.data.types.Validation;
-import ddf.catalog.validation.MetacardValidator;
-import ddf.catalog.validation.ReportingMetacardValidator;
-import ddf.catalog.validation.ValidationException;
-import ddf.catalog.validation.impl.ValidationExceptionImpl;
-import ddf.catalog.validation.impl.report.MetacardValidationReportImpl;
-import ddf.catalog.validation.impl.violation.ValidationViolationImpl;
-import ddf.catalog.validation.report.MetacardValidationReport;
-import ddf.catalog.validation.violation.ValidationViolation;
-import java.io.Serializable;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-import org.apache.commons.lang.StringUtils;
+import org.codice.ddf.validator.metacard.wkt.MetacardWktValidator;
 import org.codice.ddf.validator.wkt.WktValidator;
 
-public class MetacardLocationValidator implements MetacardValidator, ReportingMetacardValidator {
-  private static final String ERROR_MSG = Core.LOCATION + " is invalid: ";
-
-  private static final Set<String> VALIDATED_ATTRIBUTES = ImmutableSet.of(Core.LOCATION);
-
-  private final WktValidator wktValidator;
+public class MetacardLocationValidator extends MetacardWktValidator {
 
   public MetacardLocationValidator(WktValidator wktValidator) {
-    this.wktValidator = wktValidator;
-  }
-
-  @Override
-  public Optional<MetacardValidationReport> validateMetacard(Metacard metacard) {
-    MetacardValidationReportImpl report = null;
-
-    if (StringUtils.isNotEmpty(metacard.getLocation())) {
-      if (!wktValidator.isValid(metacard.getLocation())) {
-        String message = ERROR_MSG + metacard.getLocation();
-        ValidationViolation violation =
-            new ValidationViolationImpl(
-                VALIDATED_ATTRIBUTES, message, ValidationViolation.Severity.ERROR);
-        report = new MetacardValidationReportImpl();
-        report.addMetacardViolation(violation);
-      }
-    } else {
-      Attribute validationErrorAttr = metacard.getAttribute(Validation.VALIDATION_ERRORS);
-      if (validationErrorAttr != null) {
-        List<Serializable> errors = validationErrorAttr.getValues();
-        for (Serializable error : errors) {
-          String errorStr = String.valueOf(error);
-          if (errorStr.startsWith(ERROR_MSG)) {
-            ValidationViolation violation =
-                new ValidationViolationImpl(
-                    VALIDATED_ATTRIBUTES, errorStr, ValidationViolation.Severity.ERROR);
-            report = new MetacardValidationReportImpl();
-            report.addMetacardViolation(violation);
-          }
-        }
-      }
-    }
-
-    return Optional.ofNullable(report);
-  }
-
-  @Override
-  public void validate(Metacard metacard) throws ValidationException {
-    Optional<MetacardValidationReport> validationReport = validateMetacard(metacard);
-    if (validationReport.isPresent()) {
-      MetacardValidationReport report = validationReport.get();
-      Set<ValidationViolation> violations = report.getMetacardValidationViolations();
-      if (!violations.isEmpty()) {
-        metacard.setAttribute(new AttributeImpl(Core.LOCATION, (Serializable) null));
-        final ValidationExceptionImpl exception = new ValidationExceptionImpl();
-        exception.setErrors(
-            violations.stream().map(ValidationViolation::getMessage).collect(Collectors.toList()));
-        throw exception;
-      }
-    }
+    super(wktValidator, Core.LOCATION);
   }
 }
